Fix 64-bit validator to handle 66 and CS code segment prefices specially.
Review URL: http://codereview.chromium.org/464066

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@1089 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 608fd4393015d76aef4acb574aa5142b94384cc2

+++ src/trusted/validator_x86/ncopcode_desc.h
@@ -50,7 +50,7 @@
 #define MAX_BYTES_PER_X86_INSTRUCTION 15
 
 /* Defines integer to represent sets of possible opcode flags */
-typedef uint32_t OpcodeFlags;
+typedef uint64_t OpcodeFlags;
 
 /* Converts an OpcodeFlagEnum to the corresponding bit in OpcodeFlags. */
 #define InstFlag(x) (((OpcodeFlags) 1) << (x))
